To execute gl-system-install, the PATH variable must contain /home/git/bin  (ubuntu server), however a warning message is shown: "                 ***** WARNING ***** /usr/local/bin precedes /home/git/bin in your $PATH, and it *also* contains gl-setup.  This is almost certainly going to confuse you or me later.
Since gl-setup MUST be run from the PATH (and not as src/gl-setup or such),
you must fix this before running gl-setup.  The simplest way is to add

    PATH=/home/git/bin:$PATH

to the end of your bashrc or similar file.  You can even simply run that
command manually each time you log in and want to run a gitolite command."